Refactor layout code into separate files
Ref: https://github.com/ahdinosaur/blinksy/pull/55
